backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
qrcode[pil]==7.4.2
WeasyPrint
xhtml2pdf